Remove the callback that outputs the untar output
It would appear that this used to result in a silent error, where the
callable couldn't be called because it was protected. In the previous
commit we wrapped in in a Closure (we needed to since the callable
typehint was added to the exec method), fixing the "uncallable
callable". So now the output was littered with the output of the untar
command, which isn't useful info at all, and just makes the output more
difficult to read.